+++ assets/src/common/PaginationTools.jsx
@@ -8,10 +8,10 @@ function PaginationTools(props) {
     return (
         <Row>
             <Col className="justify-content-start">
-                <PerPageDropdown onChange={(limit) => props.onPerPageChange(limit)} currentLimit={props.currentLimit} limits={props.limits} />
+                <PerPageDropdown onChange={(limit) => props.onPerPageChange(limit)} currentLimit={props.currentLimit} limits={props.limits} isDisabled={props.isDisabled} />
             </Col>
             <Col className="justify-content-end d-flex">
-                <PaginationBar pages={props.totalPages} onChange={(page) => props.onPageChange(page)} currentPage={props.currentPage} showMax={props.showMaxPages} />
+                <PaginationBar pages={props.totalPages} onChange={(page) => props.onPageChange(page)} currentPage={props.currentPage} showMax={props.showMaxPages} isDisabled={props.isDisabled} />
             </Col>
         </Row>
     );
@@ -21,6 +21,7 @@ PaginationTools.propTypes = {
     currentLimit: PropTypes.number,
     currentPage: PropTypes.number,
     limits: PropTypes.arrayOf(PropTypes.number),
+    isDisabled: PropTypes.bool.isRequired,
     onPageChange: PropTypes.func.isRequired,
     onPerPageChange: PropTypes.func.isRequired,
     showMaxPages: PropTypes.number,