+++ assets/src/common/PerPageDropdown.jsx
@@ -5,7 +5,7 @@ import {Dropdown} from 'react-bootstrap';
 function PerPageDropdown(props) {
     return (
         <Dropdown onSelect={(limit) => props.onChange(limit)}>
-            <Dropdown.Toggle variant="outline-secondary">
+            <Dropdown.Toggle variant="outline-secondary" disabled={props.isDisabled}>
                 {props.currentLimit} per page
             </Dropdown.Toggle>
 
@@ -23,6 +23,7 @@ function PerPageDropdown(props) {
 PerPageDropdown.propTypes = {
     currentLimit: PropTypes.number,
     limits: PropTypes.arrayOf(PropTypes.number),
+    isDisabled: PropTypes.bool,
     onChange: PropTypes.func.isRequired
 };
 
@@ -33,7 +34,8 @@ PerPageDropdown.defaultProps = {
         50,
         100,
         250
-    ]
+    ],
+    isDisabled: false
 };
 
 export default PerPageDropdown;